Download Buttons Added
Now there's a download button on each game page to get the Windows version, which is the optimal version of the game, so the player can have a better experience playing them. Also added the .zip files of the games to download. Injection Dungeon .zip is empty since Git doesnt allow 100MB+ files to be uploaded.